pyproject.toml
[project]
name = "aws-cost-explorer-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.37.9",
    "botocore>=1.37.9",
    "googlesearch-python>=1.3.0",
    "jmespath>=1.0.1",
    "mcp[cli]>=1.3.0",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "tabulate>=0.9.0",
]
